requirements.txt
requests
python-dotenv
pathlib
logging
b2sdk